requirements.txt
crewai
crewai-tools
python-dotenv
duckduckgo-search
requests
beautifulsoup4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit
langchain-openai
